comment deletion

+++ Takeshot/base/views.py
@@ -1,7 +1,13 @@
 from django import forms
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.http.response import HttpResponse
 from django.shortcuts import render, redirect
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Message, Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
@@ -12,6 +18,50 @@ from .forms import RoomForm
 #    {'id':3,'name':'Frondend developer'},
 # ]
 
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        #try:
+            #user = User.objects.get(username=username)
+        #except:
+            #messages.error(request,'User doesnot exist')
+        
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or Password doesnot exist')
+
+    context = {'page':page}
+    return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            form.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during registration')
+    return render(request, 'base/login_register.html',{'form':form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else '' # used for search rooms
@@ -29,10 +79,23 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)  # pk is the number in url in the path
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        redirect('room', pk=room.id)
+
+    context = {'room': room,'room_messages':room_messages, 'participants':participants}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -45,10 +108,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)  # pk is the number in url in the path
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here')
+
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -58,10 +126,27 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('you are not allowed here')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
+
